+++ src/MicroElements.Metadata/Metadata/IMetadataProvider.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System.Runtime.CompilerServices;
-using MicroElements.Functional;
 
 namespace MicroElements.Metadata
 {
@@ -42,70 +41,6 @@ namespace MicroElements.Metadata
             }
 
             return propertyList;
-        }
-    }
-
-    /// <summary>
-    /// Provides extension methods for metadata providers.
-    /// </summary>
-    public static class MetadataExtensions
-    {
-        /// <summary>
-        /// Gets metadata of some type.
-        /// </summary>
-        /// <typeparam name="TMetadata">Metadata type.</typeparam>
-        /// <param name="metadataProvider">Metadata provider.</param>
-        /// <param name="metadataName">Optional metadata name.</param>
-        /// <param name="defaultValue">Default value to return if not metadata found.</param>
-        /// <returns>Metadata or default value if not found.</returns>
-        public static TMetadata GetMetadata<TMetadata>(this IMetadataProvider metadataProvider, string metadataName = null, TMetadata defaultValue = default)
-        {
-            metadataProvider.AssertArgumentNotNull(nameof(metadataProvider));
-
-            metadataName ??= typeof(TMetadata).FullName;
-            var metadata = metadataProvider.Metadata ?? metadataProvider.GetInstanceMetadata();
-
-            var propertyValue = metadata.GetPropertyValue<TMetadata>(Search.ByNameOrAlias(metadataName, ignoreCase: true));
-            if (propertyValue != null)
-                return propertyValue.Value;
-
-            return defaultValue;
-        }
-
-        /// <summary>
-        /// Sets metadata for item and returns the same metadataProvider for chaining.
-        /// </summary>
-        /// <typeparam name="TMetadataProvider">Metadata provider type.</typeparam>
-        /// <typeparam name="TMetadata">Metadata type.</typeparam>
-        /// <param name="metadataProvider">Target metadata provider.</param>
-        /// <param name="data">Metadata to set.</param>
-        /// <returns>The same metadataProvider.</returns>
-        public static TMetadataProvider SetMetadata<TMetadataProvider, TMetadata>(this TMetadataProvider metadataProvider, TMetadata data)
-            where TMetadataProvider : IMetadataProvider
-        {
-            return metadataProvider.SetMetadata(typeof(TMetadata).FullName, data);
-        }
-
-        /// <summary>
-        /// Sets metadata for item and returns the same metadataProvider for chaining.
-        /// </summary>
-        /// <typeparam name="TMetadataProvider">Metadata provider type.</typeparam>
-        /// <typeparam name="TMetadata">Metadata type.</typeparam>
-        /// <param name="metadataProvider">Target metadata provider.</param>
-        /// <param name="metadataName">Metadata name.</param>
-        /// <param name="data">Metadata to set.</param>
-        /// <returns>The same metadataProvider.</returns>
-        public static TMetadataProvider SetMetadata<TMetadataProvider, TMetadata>(this TMetadataProvider metadataProvider, string metadataName, TMetadata data)
-            where TMetadataProvider : IMetadataProvider
-        {
-            metadataName ??= typeof(TMetadata).FullName;
-            var metadata = metadataProvider.Metadata ?? metadataProvider.GetInstanceMetadata();
-            if (metadata is IMutablePropertyContainer mutablePropertyContainer)
-            {
-                mutablePropertyContainer.SetValue(new Property<TMetadata>(metadataName), data);
-            }
-
-            return metadataProvider;
         }
     }
 }